Added endpoints '/bytes/:n', '/stream-bytes/:n' and '/links/:n'.
'/bytes/:n' takes and optional 'seed' integer parameter, and returns n random
bytes of binary data (limited to 100KB in size).

'/stream-bytes/:n' is the same as above, but streams the data.  As well as a
'seed' parameter, it accepts a 'chunk_size' parameter, which determines how
many bytes are in each packet that is streamed.

'/links/:n' generates a page containing n links, each of which links to another
page containing those n links. The value of n is limited to 200.