Add back fake cookie and make requiring it conditional
Our digest auth endpoint will once again return a fake cookie but will
only require it if the query-string parameter `require-cookie` is
present.

Further, this fixes a bug in the digest auth implementation in httpbin
where we were only considering the path portions of the request-uri in
the specification. RFC 7230 is very clear that the request-uri includes
the query-string if it is present which means it is necessary in our
digest auth handling.