Status code for anything
Use case: test what would happen in a service oriented architecture if a service only returned a set of status codes for any incoming request, for example "504 Gateway Timeout".

With this change users are able to configure the client base endpoint to be `https://httpbin.org/status/504` and if the regular client appends something like `/api/v1/users/123` they would get `504` back as opposed to `404 Not Found` at the moment.