Retry-After support
Per RFC 6585 support returning a Retry-After header to test HTTP clients repeating requests.

MDN[1] says this can be returned in either an "absolute" format (Retry-After contains a fixed date) or "relative" format (Retry-After contains a number of seconds after which the request may be retried). Both formats are supported in this commit.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Retry-After