Make tox unit tests use the same versions as pipenv
When pushing the Retry-After changes, tox was failing on unit tests that worked on my system running from Pycharm CE. It seems that the reason for this is that the pipenv environment for the repo uses an older version of werkzeug (which in turn relies on an older version of flask). Adding the same fixed versions to tox.ini allows the tox tests to pass on my system so should pass on Travis too.