Remove requirements.txt & use multi-stage build
References #17

The docker-image has all dependencies "frozen" in place, so a
requirements.txt is not really needed.

It would be difficult (but possible) to ensure that the docker-image
uses the same `requirements.txt` as the one that was published. It adds
a considerable overhead to the build complexity and is not really
needed. Supporting multiple end-user environments (and
operating-systems) would also require *multiple* such files. And even
then it is not guaranteed to work everywhere.

Limiting the officially supported targets to "library" and "docker"
removes those complexities.